Add sum_mean_input and recursive_factorial functions
Implemented a sum_mean_input function in Rust that takes user input, adding it to a list until the user decides to stop. The function calculates and displays the sum and mean of the numbers inputted by the user. Also added a recursive factorial function, fact, that gets the factorial of a given number.